broke files into pieces appropriately
